corrigir erro de escrita

+++ GCS.cs
@@ -3734,7 +3734,7 @@ namespace JCFLIGHTGCS
             dataGridView1.Rows[GridCounter].Cells[Descricao.Index].Value = "Número de amostras para calibrar o AirSpeed";
 
             dataGridView1.Rows[dataGridView1.Rows.Add()].Cells[Parametro.Index].Value = "AirSpeed_Fator";
-            dataGridView1.Rows[GridCounter += 1].Cells[Unidade.Index].Value = "uint";
+            dataGridView1.Rows[GridCounter += 1].Cells[Unidade.Index].Value = "uint16_t";
             dataGridView1.Rows[GridCounter].Cells[Descricao.Index].Value = "Fator para converter a pressão em velocidade";
 
             dataGridView1.Rows[dataGridView1.Rows.Add()].DataGridView.EndEdit();
